requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
asn1crypto=1.2.0=py36_0
ballpark=1.4.0=pypi_0
blas=2.14=openblas
bzip2=1.0.8=h516909a_1
ca-certificates=2019.11.28=hecc5488_0
cairo=1.16.0=h18b612c_1001
certifi=2019.11.28=py36_0
cffi=1.13.2=py36h2e261b9_0
chardet=3.0.4=py36_1003
cloudpickle=1.2.2=py_0
configargparse=0.14.0=py36_0
cryptography=2.5=py36hb7f436b_1
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py36_0
cytoolz=0.10.1=py36h7b6447c_0
dask-core=2.8.1=py_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
expat=2.2.6=he6710b0_0
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=he4413a7_1000
freetype=2.9.1=h8a8886c_1
giflib=5.1.9=h516909a_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=hf484d3e_1000
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=2.4.0=h37c48d4_1
hdf5=1.10.4=nompi_h3c11f04_1106
icu=58.2=h9c2bf20_1
idna=2.8=py36_1000
imageio=2.6.1=py36_0
intel-openmp=2019.4=243
jasper=1.900.1=h07fcdf6_1006
joblib=0.14.0=py_0
jpeg=9c=h14c3975_1001
kiwisolver=1.1.0=py36he6710b0_0
lame=3.100=h14c3975_1001
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=14_openblas
liblapacke=3.8.0=14_openblas
libopenblas=0.3.7=h5ec1e0e_4
libpng=1.6.37=hbc83047_0
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h576950b_1
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
matplotlib=2.2.2=py36hb69df0a_2
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36h516909a_1
mkl_random=1.1.0=py36hb3f55d8_0
munch=2.5.0=py_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.4=py_0
ninja=1.9.0=py36hfd86e86_0
numpy=1.17.3=py36h95a1406_0
numpy-base=1.17.4=py36h2f8d375_0
olefile=0.46=py36_0
opencv=4.1.0=py36h79d2e43_1
openh264=1.8.0=hdbcaa40_1000
openssl=1.0.2t=h14c3975_0
pandas=0.25.3=py36he6710b0_0
parse=1.12.1=py_0
pcre=8.43=he6710b0_0
peterpy=1.0.1=pypi_0
pillow=6.2.1=py36h34e0f95_0
pip=19.3.1=py36_0
pixman=0.38.0=h516909a_1003
pycparser=2.19=py36_0
pyopenssl=19.0.0=py36_0
pyparsing=2.4.5=py_0
pyqt=5.6.0=py36h13b7fb3_1008
pysocks=1.7.1=py36_0
python=3.6.5=hc3d631a_2
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.6_cuda10.1.243_cudnn7.6.3_0
pytz=2019.3=py_0
pywavelets=1.1.1=py36h7b6447c_0
pyyaml=5.1.2=py36h7b6447c_0
pyzmq=18.1.1=py36h1768529_0
qt=5.6.3=h8bf5577_3
readline=7.0=h7b6447c_5
requests=2.22.0=py36_1
scikit-image=0.15.0=py36he6710b0_0
scikit-learn=0.22=py36hcdab131_0
scipy=1.3.2=py36h921218d_0
setuptools=42.0.2=py36_0
sip=4.18.1=py36hf484d3e_1000
six=1.13.0=py36_0
sqlite=3.30.1=h7b6447c_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchfile=0.1.0=py_0
torchvision=0.4.2=py36_cu101
tornado=6.0.3=py36h7b6447c_0
tqdm=4.40.0=py_0
urllib3=1.25.7=py36_0
visdom=0.1.8.9=0
websocket-client=0.56.0=py36_1
wheel=0.33.6=py36_0
x264=1!152.20180806=h14c3975_0
xmltodict=0.12.0=py_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yaml=0.1.7=h96e3832_1
zeromq=4.3.2=he1b5a44_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
